CORE-8711 Move GroupView and Properties under collaborators package
