+++ de-lib/src/main/java/org/iplantc/de/collaborators/client/gin/CollaboratorsGinModule.java
@@ -2,12 +2,12 @@ package org.iplantc.de.collaborators.client.gin;
 
 import org.iplantc.de.client.services.GroupServiceFacade;
 import org.iplantc.de.client.services.impl.GroupServiceFacadeImpl;
+import org.iplantc.de.collaborators.client.GroupView;
 import org.iplantc.de.collaborators.client.presenter.ManageCollaboratorsPresenter;
+import org.iplantc.de.collaborators.client.views.GroupViewImpl;
 import org.iplantc.de.collaborators.client.views.ManageCollaboratorsDialog;
 import org.iplantc.de.collaborators.client.views.ManageCollaboratorsView;
 import org.iplantc.de.collaborators.client.views.ManageCollaboratorsViewImpl;
-import org.iplantc.de.groups.client.GroupView;
-import org.iplantc.de.groups.client.views.GroupViewImpl;
 
 import com.google.gwt.inject.client.AbstractGinModule;
 import com.google.gwt.inject.client.assistedinject.GinFactoryModuleBuilder;
